完善管理员模块，新增登录模块

+++ admin/src/views/adminUser/list.vue
@@ -36,7 +36,7 @@
                             :title="'确定删除 “'+ record.username + '” 吗？'"
                             ok-text="确定"
                             cancel-text="取消"
-                            @confirm="() => onDelete(record._id)">
+                            @confirm="() => onDelete(record)">
                 <span class="table-btn text-error">
                   <a-icon type="delete" />
                   删除
@@ -118,11 +118,17 @@ export default {
       }
       this.$refs.editForm.modalVisible = true
     },
-    async onDelete (id) {
-      const res = await this.$http.delete(`rest/admin_users/${id}`)
-      if (res && res.data) {
-        this.$message.success('删除成功！')
-        this.fetchData()
+    async onDelete (item) {
+      const id = item._id
+      const username = item.username
+      if (username === 'admin') {
+        this.$message.warn('请不要删除管理员账号')
+      } else {
+        const res = await this.$http.delete(`rest/admin_users/${id}`)
+        if (res && res.data) {
+          this.$message.success('删除成功！')
+          this.fetchData()
+        }
       }
     },
     async fetchData () {
